Add functionality to save attachments as mail is indexed
Currently the new mail route only returns the number of messages
that were indexed in mongodb. For some reason, `res.json` hangs
anytime there is any attachments on the email, but works fine
when there isn't. I think it may have something to do with the file
buffer not being changed to JSON